requirements.txt
# Python dependencies
numpy>=1.24.0
matplotlib>=3.7.0
scipy>=1.10.0
customtkinter>=5.2.0
pillow>=10.0.0  # For image handling in the menu
pytest>=7.4.0    # For testing
black>=23.7.0    # For code formatting
flake8>=6.1.0    # For linting 